requirements.txt
asgiref==3.3.1
Django==3.2a1
django-crispy-forms==1.11.1
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Markdown==3.3.4
netaddr==0.8.0
Pillow==8.1.2
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
